refactor(ast): extract ref + naming helpers into refs.ts (#3572 1c)
Move extractRefName, childName, enumPropName, findDiscriminator, and
resolveGroupType out of utils/index.ts, and resolveRefName, isStringType, and
syncSchemaRef out of utils/ast.ts, into a dedicated utils/refs.ts module with
tests in refs.test.ts. ast.ts now imports resolveGroupType/resolveRefName from
refs.ts. utils/index.ts is now a re-export-only barrel; index.test.ts guards the
@kubb/ast/utils value surface with an export snapshot.
