+++ packages/ast/src/utils/refs.test.ts
@@ -1,0 +1,143 @@
+import { describe, expect, it } from 'vitest'
+import { createSchema } from '../nodes/schema.ts'
+import { childName, enumPropName, extractRefName, findDiscriminator, isStringType, resolveRefName, syncSchemaRef } from './refs.ts'
+
+describe('extractRefName', () => {
+  it('extracts schema name from component schema refs', () => {
+    expect(extractRefName('#/components/schemas/Order')).toBe('Order')
+  })
+
+  it('extracts name from component response refs', () => {
+    expect(extractRefName('#/components/responses/NotFound')).toBe('NotFound')
+  })
+
+  it('falls back to the full string when no slash is present', () => {
+    expect(extractRefName('SomeName')).toBe('SomeName')
+  })
+})
+
+describe('resolveRefName', () => {
+  it('extracts the name from a $ref pointer', () => {
+    const ref = createSchema({ type: 'ref', name: 'Pet', ref: '#/components/schemas/Pet' })
+
+    expect(resolveRefName(ref)).toBe('Pet')
+  })
+
+  it('falls back to node.name when ref is missing', () => {
+    const ref = createSchema({ type: 'ref', name: 'Pet' })
+
+    expect(resolveRefName(ref)).toBe('Pet')
+  })
+
+  it('returns null for non-ref nodes', () => {
+    expect(resolveRefName(createSchema({ type: 'string' }))).toBeNull()
+    expect(resolveRefName(undefined)).toBeNull()
+  })
+})
+
+describe('childName', () => {
+  it('returns null when parentName is undefined', () => {
+    expect(childName(undefined, 'params')).toBeNull()
+  })
+
+  it('returns PascalCase of `parentName + propName`', () => {
+    expect(childName('Order', 'params')).toBe('OrderParams')
+    expect(childName('Order', 'shipping_address')).toBe('OrderShippingAddress')
+  })
+})
+
+describe('enumPropName', () => {
+  it('combines parentName, propName, and enumSuffix', () => {
+    expect(enumPropName('Order', 'status', 'enum')).toBe('OrderStatusEnum')
+  })
+
+  it('works without parentName and with a custom suffix', () => {
+    expect(enumPropName(undefined, 'status', 'enum')).toBe('StatusEnum')
+    expect(enumPropName('Order', 'status', 'Type')).toBe('OrderStatusType')
+  })
+})
+
+describe('findDiscriminator', () => {
+  it('returns the mapping key for a matching discriminator ref', () => {
+    const mapping = {
+      cat: '#/components/schemas/Cat',
+      dog: '#/components/schemas/Dog',
+    }
+
+    expect(findDiscriminator(mapping, '#/components/schemas/Dog')).toBe('dog')
+  })
+
+  it.each([
+    {
+      label: 'mapping is missing',
+      mapping: undefined,
+      ref: '#/components/schemas/Dog' as string | undefined,
+    },
+    {
+      label: 'ref is missing',
+      mapping: { cat: '#/components/schemas/Cat' },
+      ref: undefined,
+    },
+    {
+      label: 'ref does not match any mapping entry',
+      mapping: { cat: '#/components/schemas/Cat' },
+      ref: '#/components/schemas/Dog',
+    },
+  ])('returns null when $label', ({ mapping, ref }) => {
+    expect(findDiscriminator(mapping, ref)).toBeNull()
+  })
+})
+
+describe('isStringType', () => {
+  it('returns true for plain string-like schema types', () => {
+    expect(isStringType(createSchema({ type: 'string' }))).toBe(true)
+    expect(isStringType(createSchema({ type: 'uuid' }))).toBe(true)
+    expect(isStringType(createSchema({ type: 'email' }))).toBe(true)
+    expect(isStringType(createSchema({ type: 'url' }))).toBe(true)
+    expect(isStringType(createSchema({ type: 'datetime' }))).toBe(true)
+  })
+
+  it('returns true for date/time with string representation', () => {
+    expect(isStringType(createSchema({ type: 'date', representation: 'string' }))).toBe(true)
+    expect(isStringType(createSchema({ type: 'time', representation: 'string' }))).toBe(true)
+  })
+
+  it('returns false for date/time with date representation and non-string scalars', () => {
+    expect(isStringType(createSchema({ type: 'date', representation: 'date' }))).toBe(false)
+    expect(isStringType(createSchema({ type: 'time', representation: 'date' }))).toBe(false)
+    expect(isStringType(createSchema({ type: 'number' }))).toBe(false)
+  })
+})
+
+describe('syncSchemaRef', () => {
+  it('returns a merged schema for a ref node that has a resolved schema', () => {
+    const resolved = createSchema({ type: 'object' })
+    const ref = createSchema({
+      type: 'ref',
+      name: 'Pet',
+      ref: '#/components/schemas/Pet',
+      schema: resolved,
+    })
+
+    const merged = syncSchemaRef(ref)
+    expect(merged).not.toBeNull()
+    expect(merged?.type).toBe('object')
+  })
+
+  it('returns a merged schema with sibling overrides applied over the resolved schema', () => {
+    const resolved = createSchema({ type: 'object', description: 'Original' })
+    const ref = createSchema({
+      type: 'ref',
+      name: 'Pet',
+      ref: '#/components/schemas/Pet',
+      schema: resolved,
+      description: 'Override',
+      readOnly: true,
+    })
+
+    const merged = syncSchemaRef(ref)
+    expect(merged?.description).toBe('Override')
+    expect(merged?.readOnly).toBe(true)
+    expect(merged?.type).toBe('object')
+  })
+})
